Update tests to not require db
