+++ packages/partner/tests/php/test_Partner.php
@@ -1,20 +1,23 @@
 <?php
 
 namespace Automattic\Jetpack;
+use PHPUnit\Framework\TestCase;
 
-class Partner_Test extends \WP_UnitTestCase {
+use Brain\Monkey;
+use Brain\Monkey\Functions;
+use Brain\Monkey\Filters;
+
+class Partner_Test extends TestCase {
 
 	const TEST_CODE = 'abc-123';
 
 	public function tearDown() {
-		if ( has_filter( 'jetpack_partner_subsidiary_id', array( $this, 'partner_code_filter' ) ) ) {
-			remove_filter( 'jetpack_partner_subsidiary_id', array( $this, 'partner_code_filter' ) );
-		}
 		parent::tearDown();
 	}
 
 	public function setUp() {
 		parent::setUp();
+		Monkey\setUp();
 		Partner::reset();
 	}
 
@@ -45,8 +48,11 @@ class Partner_Test extends \WP_UnitTestCase {
 	 * @param string $code_type Partner code type.
 	 * @param string $option_name Option and filter name.
 	 * @param string $query_string_name Query string variable name.
+	 *
+	 * @throws Monkey\Expectation\Exception\ExpectationArgsRequired
 	 */
 	public function test_partner_code_is_empty_by_defalt( $code_type, $option_name, $query_string_name ) {
+		Functions\expect( 'get_option' )->once()->with( $option_name )->andReturn( '' );
 		$this->assertEmpty( Partner::init()->get_partner_code( $code_type ) );
 	}
 
@@ -56,9 +62,11 @@ class Partner_Test extends \WP_UnitTestCase {
 	 * @param string $code_type Partner code type.
 	 * @param string $option_name Option and filter name.
 	 * @param string $query_string_name Query string variable name.
+	 *
+	 * @throws Monkey\Expectation\Exception\ExpectationArgsRequired
 	 */
 	public function test_partner_code_is_set_via_option( $code_type, $option_name, $query_string_name ) {
-		add_option( $option_name, self::TEST_CODE );
+		Functions\expect( 'get_option' )->once()->with( $option_name, '' )->andReturn( self::TEST_CODE );
 		$this->assertEquals( self::TEST_CODE, Partner::init()->get_partner_code( $code_type ) );
 	}
 
@@ -68,33 +76,12 @@ class Partner_Test extends \WP_UnitTestCase {
 	 * @param string $code_type Partner code type.
 	 * @param string $option_name Option and filter name.
 	 * @param string $query_string_name Query string variable name.
+	 *
+	 * @throws Monkey\Expectation\Exception\ExpectationArgsRequired
 	 */
 	public function test_partner_code_is_set_via_filter( $code_type, $option_name, $query_string_name ) {
-		add_filter( $option_name, array( $this, 'partner_code_filter' ) );
+		Functions\expect( 'get_option' )->once()->with( $option_name)->andReturn( '' );
+		Filters\expectApplied( $option_name )->once()->with('')->with( self::TEST_CODE );
 		$this->assertEquals( self::TEST_CODE, Partner::init()->get_partner_code( $code_type ) );
-	}
-
-	/**
-	 * @dataProvider code_provider
-	 *
-	 * @param string $code_type Partner code type.
-	 * @param string $option_name Option and filter name.
-	 * @param string $query_string_name Query string variable name.
-	 */
-	public function test_partner_codes_are_added_to_connect_url( $code_type, $option_name, $query_string_name ) {
-		Partner::init();
-		add_filter( $option_name, array( $this, 'partner_code_filter' ) );
-		$jetpack = \Jetpack::init();
-		$url     = $jetpack->build_connect_url();
-
-		$parsed_vars = array();
-		parse_str( wp_parse_url( $url, PHP_URL_QUERY ), $parsed_vars );
-
-		$this->assertArrayHasKey( $query_string_name, $parsed_vars );
-		$this->assertSame( self::TEST_CODE, $parsed_vars[ $query_string_name ] );
-	}
-
-	public function partner_code_filter() {
-		return self::TEST_CODE;
 	}
 }
